Restructures Code
The primary changes included creating subfolders and renaming current folders

- `Entities` is now `Models` in which requests, responses, and classes related to DynamoDB are subdivided inside
- `Repositories` is now called `Services`
