+++ src/CloudLog-API/Controllers/V1/UserInfoController.cs
@@ -1,7 +1,8 @@
 using Asp.Versioning;
-using CloudLogAPI.Entities;
 using CloudLogAPI.Exceptions;
-using CloudLogAPI.Repositories;
+using CloudLogAPI.Models.Requests;
+using CloudLogAPI.Models.Responses;
+using CloudLogAPI.Services;
 using Microsoft.AspNetCore.Authorization;
 using Microsoft.AspNetCore.Mvc;
 using Microsoft.IdentityModel.Tokens;
