--- src/CloudLog-API/Entities/LogJumpRequest.cs
@@ -1,10 +1,0 @@
-using System.ComponentModel.DataAnnotations;
-using System.Text.Json.Serialization;
-using CloudLogAPI.Records;
-
-namespace CloudLogAPI.Entities;
-
-public sealed class LogJumpRequest
-{
-    public LoggedJump? Jump { get; init; }
-}
