+++ src/CloudLog-API/Models/DynamoDB/JumpType.cs
@@ -1,6 +1,4 @@
-using System.Text.Json.Serialization;
-
-namespace CloudLogAPI.Entities;
+namespace CloudLogAPI.Models.DynamoDB;
 
 [Flags]
 public enum JumpType
