+++ src/CloudLog-API/Models/DynamoDB/UserInfo.cs
@@ -2,7 +2,7 @@ using System.ComponentModel.DataAnnotations;
 using System.Text.Json.Serialization;
 using Amazon.DynamoDBv2.DataModel;
 
-namespace CloudLogAPI.Records;
+namespace CloudLogAPI.Models.DynamoDB;
 
 [DynamoDBTable(nameof(UserInfo))]
 public sealed class UserInfo
