+++ src/CloudLog-API/Models/Requests/EditJumpRequest.cs
@@ -1,6 +1,6 @@
-using CloudLogAPI.Records;
+using CloudLogAPI.Models.DynamoDB;
 
-namespace CloudLogAPI.Entities;
+namespace CloudLogAPI.Models.Requests;
 
 public sealed class EditJumpRequest
 {
