+++ src/CloudLog-API/Models/Requests/LogJumpRequest.cs
@@ -1,0 +1,8 @@
+using CloudLogAPI.Models.DynamoDB;
+
+namespace CloudLogAPI.Models.Requests;
+
+public sealed class LogJumpRequest
+{
+    public LoggedJump? Jump { get; init; }
+}
